favorites: use portable conventions
Previously, `loadFavorites` could get an empty string for `$HOME`.
This is not treated as an error, so the favorites list returns with 1
item, the root.
`newFavoritesUI` is hard coded to expect at least 2 elements, and will
panic otherwise.

Now, failing to load a home directory results in an error up front.
Delimiters are also swapped out here to be their platform native rune.