all: add portable path delimiter logic
This replaces some logic where paths are expected to be `/` rooted and
`/` delimited.

On Windows this creates an oddity around the favorite entry for `$HOME`.
The home directory is absolute, while all other paths are relative to
the system volume.
I.e. `C:\Users\...` will not match against `\Users\...` when considering
duplicate entries into the favorites list.

Broader changes to pathing logic would need to be changed for this.
Specifically in regard to volume roots.